move <Loader /> to components director
