+++ app/components/Avatar/index.js
