+++ app/containers/Articles/tests/index.test.js
@@ -4,8 +4,8 @@ import { shallow, mount } from 'enzyme';
 import ReactPaginate from 'react-paginate';
 import { Alert, Card, CardTitle, CardText, Badge, CardLink, Row, Col, Button } from 'reactstrap';
 import Loader from 'components/Loader';
+import Avatar from 'components/Avatar';
 import Wrapper from '../Wrapper';
-import Avatar from '../Avatar';
 import { Articles, mapDispatchToProps } from '../index';
 import { fetchArticles } from '../actions';
 import { BASE_LIMIT } from '../../../utils/url';
